Implemented GameRules for cost multipliers.
Two new gamerules are added: Vocal Cooldown Multiplier and Tome Cooldown Multiplier. Both of these are percentage multipliers on the final cost of the spell before the cooldown/hunger/damage is applied. It is entirely separate from the thresholds/maximum cost gamerules. One has no effect on the other. Additionally, the Hunger effect duration now scales with how far the cost is over the threshold.